Module steel_cent::currency

Besides defining the Currency struct, this module provides constants for most real-world currencies as well as functions to look up a Currency by alphabetic or numeric code.

Structs

Currency
A monetary currency with text and numeric codes and a standard number of decimal places, as per ISO 4217.

Constants

AED

AFN

ALL

AMD

ANG

AOA

ARS

AUD

AWG

AZN

BAM

BBD

BDT

BGN

BHD

BIF

BMD

BND

BOB

BRL

BSD

BTN

BWP

BYN

BZD

CAD

CDF

CHF

CLP

CNY

COP

CRC

CUP

CVE

CZK

DJF

DKK

DOP

DZD

EGP

ERN

ETB

EUR

FJD

FKP

GBP

GEL

GHS

GIP

GMD

GNF

GTQ

GYD

HKD

HNL

HRK

HTG

HUF

IDR

ILS

INR

IQD

IRR

ISK

JMD

JOD

JPY

KES

KGS

KHR

KMF

KPW

KRW

KWD

KYD

KZT

LAK

LBP

LKR

LRD

LSL

LYD

MAD

MDL

MGA

MKD

MMK

MNT

MOP

MRU

MUR

MVR

MWK

MXN

MYR

MZN

NAD

NGN

NIO

NOK

NPR

NZD

OMR

PAB

PEN

PGK

PHP

PKR

PLN

PYG

QAR

RON

RSD

RUB

RUR

RWF

SAR

SBD

SCR

SDG

SEK

SGD

SHP

SLL

SOS

SRD

SSP

STN

SVC

SYP

SZL

THB

TJS

TMT

TND

TOP

TRY

TTD

TWD

TZS

UAH

UGX

USD

UYU

UZS

VEF

VND

VUV

WST

XAF

XCD

XOF

XPF

YER

ZAR

ZMW

ZWL

Functions

with_code
Looks up a "standard" currency by its three-letter code.

with_numeric_code
Looks up a "standard" currency by its numeric code.