Struct steel_cent::SmallMoney

pub struct SmallMoney {
    pub currency: Currency,
    // some fields omitted
}

A signed amount of money in a certain currency, with the currency's standard number of decimal places.

The range of supported values is small enough to be inappropriate for some practical applications. See max and min below.

Note that the arithmetic ops implementations all delegate to the checked_ methods, panicking on None, even in a release build. (This is in contrast to primitive ops, which will overflow in a release build.)

A SmallMoney is 64 bits in size.

assert_eq!(8, std::mem::size_of::<SmallMoney>());

Examples

let price = SmallMoney::of_major_minor(USD, 19, 95);
let shipping_and_handling = SmallMoney::of_major(USD, 10);
let convenience_charge = SmallMoney::of_major(USD, 6);
let discount: f64 = 1.0 - 0.2; // 20% off
let discounted_price = price * discount;
let fees = shipping_and_handling + convenience_charge;
let total = discounted_price + fees;
println!("price: {:?}, discounted_price: {:?}", price, discounted_price);
assert_eq!(SmallMoney::of_minor(USD, 1596), discounted_price);
assert_eq!(SmallMoney::of_minor(USD, 3196), total);
assert_eq!((price * discount) + shipping_and_handling + convenience_charge, total);

Fields
currency: Currency
Implementations

impl SmallMoney

pub fn of_minor(currency: Currency, amount_minor: i32) -> Self

Creates a SmallMoney from its "minor" unit (e.g. US cents to USD).

pub fn of_major(currency: Currency, amount_major: i32) -> Self

Creates a SmallMoney from its "major" unit (e.g. US dollars to USD).

pub fn of_major_minor(
currency: Currency,
amount_major: i32,
amount_minor: i32
) -> Self

pub fn zero(currency: Currency) -> Self

pub fn min(currency: Currency) -> Self

assert_eq!(SmallMoney::of_major_minor(USD, -21_474_836, -48), SmallMoney::min(USD));

pub fn max(currency: Currency) -> Self

assert_eq!(SmallMoney::of_major_minor(USD, 21_474_836, 47), SmallMoney::max(USD));

pub fn major_part(&self) -> i32

Returns the major unit part of the amount.

assert_eq!(2, SmallMoney::of_minor(USD, 2_99).major_part());

pub fn minor_part(&self) -> i32

Returns the minor unit part of the amount.

assert_eq!(99, SmallMoney::of_minor(USD, 2_99).minor_part());

pub fn minor_amount(&self) -> i32

Returns the total amount in the currency's minor unit.

let two_dollars = SmallMoney::of_major(USD, 2);
assert_eq!(200, two_dollars.minor_amount());

pub fn convert_to(&self, currency: Currency, conversion_multiplier: f64) -> Self

let five_usd = SmallMoney::of_major(USD, 5);
let three_pound_seventy_five = SmallMoney::of_major_minor(GBP, 3, 75);
let gbp_per_usd = 0.75;
let five_eleven_yen = SmallMoney::of_major(JPY, 511);
let jpy_per_usd = 102.15;
assert_eq!(three_pound_seventy_five, five_usd.convert_to(GBP, gbp_per_usd));
assert_eq!(five_usd, three_pound_seventy_five.convert_to(USD, gbp_per_usd.recip()));
assert_eq!(510.75, 5.0 * jpy_per_usd); // but JPY has zero decimal places.
assert_eq!(five_eleven_yen, five_usd.convert_to(JPY, jpy_per_usd)); // rounded.
assert_eq!(five_usd, five_eleven_yen.convert_to(USD, jpy_per_usd.recip())); // rounded.

pub fn checked_abs(&self) -> Option<Self>

Returns absolute value, except for the minimum value, which cannot be negated.

let c = SmallMoney::of_major(USD, 100);
assert_eq!(Some(c), c.checked_abs());
assert_eq!(Some(c), (-c).checked_abs());
assert_eq!(None, SmallMoney::min(USD).checked_abs());

pub fn abs(&self) -> Self

Returns absolute value, except for the minimum value, which cannot be negated.

let c = SmallMoney::of_major(USD, 100);
assert_eq!(c, c.abs());
assert_eq!(c, (-c).abs());

Panics

Panics for the minimum value.

pub fn checked_add(self, other: Self) -> Option<Self>

assert_eq!(Some(SmallMoney::of_minor(USD, 4_01)),
           SmallMoney::of_major(USD, 4).checked_add(SmallMoney::of_minor(USD, 1)));
assert_eq!(None, SmallMoney::max(USD).checked_add(SmallMoney::of_minor(USD, 1)));

Panics

Panics when currencies differ.

pub fn checked_sub(self, other: Self) -> Option<Self>

assert_eq!(Some(SmallMoney::of_minor(USD, 3_99)),
           SmallMoney::of_major(USD, 4).checked_sub(SmallMoney::of_minor(USD, 1)));
assert_eq!(None, SmallMoney::min(USD).checked_sub(SmallMoney::of_minor(USD, 1)));

Panics

Panics when currencies differ.

pub fn checked_mul(self, n: i32) -> Option<Self>

assert_eq!(Some(SmallMoney::of_major(USD, 8)), SmallMoney::of_major(USD, 4).checked_mul(2));
assert_eq!(None, SmallMoney::max(USD).checked_mul(2));
assert_eq!(None, SmallMoney::min(USD).checked_mul(-1));

pub fn checked_mul_f(self, n: f64) -> Option<Self>

Checked multiplication by a float with rounding. Note that a float has less-than-integer precision for very large and very small amounts of money, which can result in surprising rounding errors.

assert_eq!(Some(SmallMoney::of_minor(USD, 8_40)), SmallMoney::of_major(USD, 4).checked_mul_f(2.1));
assert_eq!(None, SmallMoney::max(USD).checked_mul_f(1.01));
assert_eq!(None, SmallMoney::min(USD).checked_mul_f(-1.0));

pub fn checked_div(self, n: i32) -> Option<Self>

assert_eq!(Some(SmallMoney::of_major(USD, 2)), SmallMoney::of_minor(USD, 4_01).checked_div(2));
assert_eq!(None, SmallMoney::of_major(USD, 1).checked_div(0));
assert_eq!(None, SmallMoney::min(USD).checked_div(-1));

pub fn checked_rem(self, n: i32) -> Option<Self>

assert_eq!(Some(SmallMoney::of_minor(USD, 1)), SmallMoney::of_minor(USD, 4_01).checked_rem(2));
assert_eq!(None, SmallMoney::of_major(USD, 1).checked_rem(0));
assert_eq!(None, SmallMoney::min(USD).checked_rem(-1));

pub fn checked_neg(self) -> Option<Self>

Negates the value, except for the minimum value, which cannot be negated.

assert_eq!(Some(SmallMoney::of_major(USD, -1)), SmallMoney::of_major(USD, 1).checked_neg());
assert_eq!(None, SmallMoney::min(USD).checked_neg());

pub fn saturating_add(self, other: Self) -> Self

assert_eq!(SmallMoney::of_minor(USD, 4_01),
           SmallMoney::of_major(USD, 4).saturating_add(SmallMoney::of_minor(USD, 1)));
assert_eq!(SmallMoney::max(USD),
           SmallMoney::max(USD).saturating_add(SmallMoney::of_minor(USD, 1)));

Panics

Panics when currencies differ.

pub fn saturating_sub(self, other: Self) -> Self

assert_eq!(SmallMoney::of_minor(USD, 3_99),
           SmallMoney::of_major(USD, 4).saturating_sub(SmallMoney::of_minor(USD, 1)));
assert_eq!(SmallMoney::min(USD),
           SmallMoney::min(USD).saturating_sub(SmallMoney::of_minor(USD, 1)));

Panics

Panics when currencies differ.

pub fn saturating_mul(self, n: i32) -> Self

assert_eq!(SmallMoney::of_major(USD, 8), SmallMoney::of_major(USD, 4).saturating_mul(2));
assert_eq!(SmallMoney::max(USD), SmallMoney::max(USD).saturating_mul(2));
assert_eq!(SmallMoney::max(USD), SmallMoney::min(USD).saturating_mul(-1));